+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Added
+- Beta: Compact visualization for table (for usage in panorama)
+- aggregate data also without value column
+
 ## 5.3.0
 ### Fixed
 - Reading the third level in a json array #434